requirements.txt
aiohttp>=3.9
aiohttp-jinja2>=1.6
aiofiles>=23.0
python-dotenv>=1.0
